# Corrected CH2 & CH2* species for reactions in JetSurF2.0