This Pull Request is submitted with reference to the email communication with Prof. Alon Grinberg Dana on issue #649 .
The JetSurF2.0 reaction library is corrected for the below reactions:
CH2 + O <=> HCO + H 
CH2 + H2 <=> H + CH3
CH3 + OH <=> CH2 + H2O
CH3 + CH2 <=> C2H4 + H
CH4 + CH2 <=> CH3 + CH3
C2H4 + CH2 <=> aC3H5 + H
MultiArrhenius kinetic type is removed from the above reactions as per the JetSurf2.0 mechanism available online (https://web.stanford.edu/group/haiwanglab/JetSurF/JetSurF2.0/Mech_JetSurF2.0.txt)

Furthermore, the following reactions are added to the library based on the web version of the mechanism:
CH3 + OH <=> CH2* + H2O
CH2* + H2 <=> CH3 + H
CH3 + CH2* <=> C2H4 + H
C2H4 + CH2* <=> aC3H5 + H
CH4 + CH2* <=> CH3 + CH3
CH2* + O <=> HCO + H

@alongd Prof. Alon Grinberg Dana, I kindly request you to review the changes and provide your feedback on the same.