@alongd 2026-08-31
Apologies for the long silence on this one, @PoojaNem — thank you for the contribution, and for the diagnosis, which is correct and worth spelling out for the record.

## The bug

When JetSurF2.0 was imported from ChemKin into RMG, every reaction that the source mechanism gives *separately* for triplet `CH2` and singlet `CH2*` — same co-reactant, same products — was collapsed onto the **triplet** as a `MultiArrhenius` with both rate expressions, and the singlet entry was dropped. Two things go wrong at once: the triplet runs at `k(triplet) + k(singlet)`, and the singlet channel disappears entirely.

For `CH3 + OH <=> CH2 + H2O` that is not a small error:

| T | k(triplet), source | k in RMG's copy | inflation |
|---|---|---|---|
| 1000 K | 2.31e11 | 2.52e13 | 109× |
| 1500 K | 1.10e12 | 2.61e13 | 24× |
| 2000 K | 2.74e12 | 2.78e13 | 10× |

(cm³/(mol·s); worst exactly where a laminar flame speed is most sensitive to it.) That is enough to account for the discrepancy reported in #649.

The tell that this is an import artifact rather than a real duplicate pair: `Mech_JetSurF2.0.txt` uses the `DUPLICATE` keyword for the HO2/H2O2/CO pairs **only**. None of these eight methylene reactions carries it in the source — they are distinct species, not duplicates. The `duplicate = True` flags left on some of the RMG entries are therefore stale, and are removed here.

## What I pushed on top of your branch

Your commit fixed 6 of the 8 affected pairs. I've pushed `b3414f2` onto this branch, which:

- **adds the 2 pairs that were missed** — `CH2/CH2* + OH <=> CH2O + H` (2.0e13 / 3.0e13, GRI) and `C2H2 + CH2/CH2* <=> C3H3 + H` (1.2e13 Ea=6620 / 2.0e13, 97WAN/FRE);
- **strips the remaining stale `duplicate = True`** flags in the methylene set (entries 41, 61, and the pairs above);
- cosmetics only beyond that: replaces the `Originally from reaction library: Unclassified` boilerplate with the source's own `shortDesc`, normalizes `degeneracy = 1.0` → `1`, restores the file's comma spacing.

All 16 rate expressions (8 pairs × 2 spin states) were checked line-by-line against `Mech_JetSurF2.0.txt`. Net: +52 / −97 lines, `entry(...)` data only.

## Validation

Loaded the library in `rmg_env` before and after, and asserted for each of the 16 channels that it exists exactly once, carries a plain `Arrhenius` rather than a `MultiArrhenius`, matches the source A factor, and involves a methylene of the right multiplicity:

|  | `main` (fc7bb138) | this branch (b3414f2) |
|---|---|---|
| entries | 2087 | 2095 |
| `check_for_duplicates` | passes | passes |
| channels correct | **0 / 16** | **16 / 16** |

The 14 `Only Arrhenius and PDepArrhenius kinetics supported for duplicate reactions` warnings emitted at load are **identical before and after** — pre-existing and unrelated to this change.

## CI

Being straight about this rather than claiming green: the checks here are red, but not because of this PR. In `Conda Build`, Python 3.7/3.8/3.9/3.10/3.12 all pass and only 3.11 fails, at the `Setup Miniforge` step — before anything is built. `Continuous Integration` fails at the same `Setup Miniforge Python 3.11` step, never reaching the database tests. Independently, `main`'s own scheduled CI has failed every day this week (latest: `Install and link Julia dependencies`). The 2024 regression-test comment above is likewise unrelated — its diffs are all in `intra_H_migration`, `Peroxyl_Disproportionation`, `H_Abstraction` and the sulfur library; none of those test cases loads JetSurF2.0.

(The runs on this PR had also been sitting at `action_required` since 2024 — the fork-contributor approval gate — which is why no CI ever reported. Approved and re-run now.)

## Out of scope, for a follow-up

While auditing I noticed `C4H81 + O`, `C5H6 + O` and `C5H5 + OH` also carry `MultiArrhenius` in this library while carrying no `DUPLICATE` in the source. Those may be the same import artifact or may be legitimate isomer collapses — it needs checking, but deliberately not in this PR. A full audit of all ~2100 JetSurF2.0 reactions against the source is worth filing separately.

Once CI is sorted this is good to merge from my side, and #649 can be closed with it.