@alongd 2026-08-31
**CI update — correcting my note above.** Both workflows had been sitting on stale 2024 runs; now that they have actually re-run against `b3414f2`:

- **`Conda Build` is fully green** — all six Python versions (3.7–3.12) build, including 3.11. The `Setup Miniforge` failure I described was stale, not real.
- **`Continuous Integration` fails at `Install and link Julia dependencies`** — which is the *exact same job and step* that `main`'s own scheduled CI has been failing on all week (latest run on `main`: `build-and-test-linux` → `Install and link Julia dependencies`). It never reaches `Run Database Tests`.

So the one remaining red check is an environment breakage that `main` shares, independent of this PR. Nothing here fails on the change itself.