@alongd 2026-08-31
## Cantera before/after

Quantified the effect on n-decane/air, since #649 reported a flame-speed difference without a magnitude.

**Method, because it bounds what these numbers mean.** Both mechanisms are exported from RMG-database itself — the JetSurF2.0 kinetics library plus the matching JetSurF2.0 thermo library, which covers all 341 species — rather than from the Stanford files. The database carries no JetSurF transport data, so transport is RMG-estimated. Both sides get byte-identical treatment and differ *only* in the eight methylene channels, so the deltas below are meaningful; the absolute values should not be read against experiment.

Checks that the two exports differ only where intended:

| | before | after |
|---|---|---|
| species / reactions | 341 / 2102 | 341 / 2102 |
| `CH2` appears in | 36 reactions | 28 reactions |
| `CH2*` appears in | 16 reactions | **24 reactions** (+8) |
| `DUPLICATE`-flagged | 45 | 29 (−16 = 8 pairs × 2) |

Equal reaction totals are expected, not suspicious: each `MultiArrhenius` expands to a `DUPLICATE` pair on export, so 2087 + 15 = 2095 + 7 = 2102.

### Ignition delay (φ = 1.0, max dT/dt)

| P (atm) | T (K) | before (ms) | after (ms) | change |
|---|---|---|---|---|
| 1 | 1000 | 32.17 | 32.32 | +0.44% |
| 1 | 1200 | 1.455 | 1.559 | +7.12% |
| 1 | 1400 | 0.1219 | 0.1411 | **+15.75%** |
| 20 | 1000 | 2.245 | 2.246 | +0.05% |
| 20 | 1200 | 0.1484 | 0.1499 | +0.96% |
| 20 | 1400 | 0.01655 | 0.01716 | +3.65% |

### Laminar flame speed (Tu = 400 K, 1 atm, mixture-averaged)

| φ | before (cm/s) | after (cm/s) | change |
|---|---|---|---|
| 0.7 | 40.74 | 39.79 | −2.33% |
| 1.0 | 61.26 | 60.58 | −1.11% |
| 1.3 | 46.03 | 47.14 | +2.42% |

### Reading this — and a correction to my comment above

I wrote earlier that the rate inflation was "enough to account for the discrepancy reported in #649." That was asserted before I had measured it, and the measurement makes it more specific than that: the **flame-speed** effect is real but modest, 1–2%, and it changes sign across equivalence ratio (slower when lean, faster when rich). The **autoignition** effect is the larger one, up to +16% at 1 atm / 1400 K, which is where the inflated `CH3 + OH` channel carries the most weight. So @PoojaNem's fix matters more for ignition than for flame speed, and a user seeing a large flame-speed shift from this library alone may have a second cause worth chasing.

None of that changes whether the fix is correct — it plainly is, and the independent corroboration is in this repo already: **JetSurF1.0 imported the same source chemistry correctly**, carrying `CH3 + OH <=> CH2 + H2O` and `CH3 + OH <=> CH2(S) + H2O` as separate entries. The defect is specific to the 2.0 import, exactly as #649 suspected.